changed signature of BasicStatistics methods -- they now act on Number instead of Double or Float

+++ src/org/fhcrc/cpl/viewer/amt/AmtDatabaseMatcher.java
@@ -426,7 +426,7 @@ public class AmtDatabaseMatcher
                 for (Feature feature : featuresThisModState)
                      retentionTimesThisModState.add(feature.getTime());
 
-                float medianTime = BasicStatistics.median(retentionTimesThisModState);
+                float medianTime = (float) BasicStatistics.median(retentionTimesThisModState);
 
                 int lastScan = 0;
                 Feature firstFeature = null;
